feat: rename MealBuilder to MealAdd and refactor MealAddScreen layout
